+++ src/TestSprint1/KilosTest.java
@@ -1,4 +1,4 @@
-package Test;
+package TestSprint1;
 
 import org.junit.Assert;
 import org.junit.Test;
