+++ src/TestSprint1/PoundsTest.java
@@ -1,90 +1,87 @@
-package Test;
+package TestSprint1;
 
 import org.junit.Assert;
 import org.junit.Test;
 
-import java.util.Scanner;
-
-public class GramsTest {
+public class PoundsTest {
     // 1.positive number converting to kg/gr/ounces /done/
     // 3.negative numbers converting to kg/gr/ounces /done/
     // 4.number zero converting to kg/gr/ounces /done/
 
     @Test
     public void verifyPositiveNumberToKg(){
-        double expectedResult = 0.01;
+        double expectedResult = 4.53592;
         double a = 10;
-        double actualResult = a/1000;
+        double actualResult = a*0.453592;
         String message = "expected = "+ expectedResult + " : Actual = "+actualResult;
         Assert.assertTrue(message, expectedResult==actualResult);
     }
     @Test
-    public void verifyPositiveNumberToOz(){
-        double expectedResult = 283.4949254408346;
+    public void verifyPositiveNumberToGr(){
+        double expectedResult = 4535.92;
         double a = 10;
-        double actualResult = a/0.035274;
+        double actualResult = a*453.592;
         String message = "expected = "+ expectedResult + " : Actual = "+actualResult;
         Assert.assertTrue(message, expectedResult==actualResult);
     }
     //pounds = grams * 0.00220462;
     @Test
-    public void verifyPositiveNumberToPounds(){
-        double expectedResult = 4535.929094356397;
+    public void verifyPositiveNumberToOz(){
+        double expectedResult = 160;
         double a = 10;
-        double actualResult = a/0.00220462;
+        double actualResult = a*16.0;
         String message = "expected = "+ expectedResult + " : Actual = "+actualResult;
         Assert.assertTrue(message, expectedResult==actualResult);
     }
     @Test
     public void verifyNegativeNumberToKg(){
-        double expectedResult = -0.01;
+        double expectedResult = -4.53592;
         double a = -10;
-        double actualResult = a/1000;
+        double actualResult = a*0.453592;
         String message = "expected = "+ expectedResult + " : Actual = "+actualResult;
         Assert.assertTrue(message, expectedResult==actualResult);
     }
     @Test
-    public void verifyNegativeNumberToOz(){
-        double expectedResult = -283.4949254408346;
+    public void verifyNegativeNumberToGr(){
+        double expectedResult = -4535.92;
         double a = -10;
-        double actualResult = a/0.035274;
+        double actualResult = a*453.592;
         String message = "expected = "+ expectedResult + " : Actual = "+actualResult;
         Assert.assertTrue(message, expectedResult==actualResult);
     }
     //pounds = grams * 0.00220462;
     @Test
-    public void verifyNegativeNumberToPounds(){
-        double expectedResult = -4535.929094356397;
+    public void verifyNegativeNumberToOz(){
+        double expectedResult = -160;
         double a = -10;
-        double actualResult = a/0.00220462;
+        double actualResult = a*16.0;
         String message = "expected = "+ expectedResult + " : Actual = "+actualResult;
         Assert.assertTrue(message, expectedResult==actualResult);
     }
 
-    @Test (expected =ArithmeticException.class)
+    @Test (expected = AssertionError.class)
     public void verifyZeroToKg() {
         int expectedResult = 0;
         int a = 0;
-        double actualResult = 1000/a;
+        double actualResult = 0.453592/a;
         String message = "expected=" + expectedResult + " : Actual=" + actualResult;
         Assert.assertTrue(message, expectedResult == actualResult);
+    }
 
+    @Test (expected = AssertionError.class)
+    public void verifyZeroToGr() {
+        int expectedResult = 0;
+        int a = 0;
+        double actualResult = 453.592/a;
+        String message = "expected=" + expectedResult + " : Actual=" + actualResult;
+        Assert.assertTrue(message, expectedResult == actualResult);
     }
 
     @Test (expected = AssertionError.class)
     public void verifyZeroToOz() {
         int expectedResult = 0;
         int a = 0;
-        double actualResult = 0.035274/a;
-        String message = "expected=" + expectedResult + " : Actual=" + actualResult;
-        Assert.assertTrue(message, expectedResult == actualResult);
-    }
-
-    @Test (expected = AssertionError.class)
-    public void verifyZeroToPound() {
-        int expectedResult = 0;
-        int a = 0;
-        double actualResult = 0.00220462 /a;
+        double actualResult = 16.0 /a;
         String message = "expected=" + expectedResult + " : Actual=" + actualResult;
         Assert.assertTrue(message, expectedResult == actualResult);
     }
