+++ src/sprint2/main/FahrenheitToCelsius.java
@@ -1,0 +1,8 @@
+package sprint2.main;
+
+public class FahrenheitToCelsius {
+    double result;
+    FahrenheitToCelsius(double num1){
+        result = (num1-32)*5/9;
+    }
+}
