+++ src/sprint2/main/FahrenheitToKelvin.java
@@ -1,0 +1,9 @@
+package sprint2.main;
+
+public class FahrenheitToKelvin {
+    double result;
+    FahrenheitToKelvin(double num1){
+        result = (num1-32)*5/9+273.15;
+    }
+
+}
