--- src/sprint2/main/constructors.java
@@ -1,7 +1,0 @@
-package sprint2.main;
-
-public class constructors {
-
-
-
-}
