--- src/sprint2/main/mainClassConverter.java
@@ -1,8 +1,0 @@
-package sprint2.main;
-
-public class mainClassConverter {
-    public static void main(String[] args) {
-
-
-    }
-}
